+++ arcade_machine/images/image.py
@@ -1,0 +1,7 @@
+
+class Image:
+    def __init__(self, file_path):
+        self.file_path = file_path
+
+    def get_file_path(self):
+        return self.file_path