+++ arcade_machine/images/thumbnail_images.py
@@ -1,0 +1,8 @@
+from arcade_machine.images.image import Image
+
+ASTEROID_UNIVERSE_THUMBNAIL = Image('arcade_machine/resources/images/thumbnails/Asteroid.png')
+MARS_LANDER_THUMBNAIL = Image('arcade_machine/resources/images/thumbnails/Mars_Lander.png')
+PONG_THUMBNAIL = Image('arcade_machine/resources/images/thumbnails/No_Image.png')
+SNAKE_THUMBNAIL = Image('arcade_machine/resources/images/thumbnails/No_Image.png')
+
+no_thumbnail = Image('arcade_machine/resources/images/thumbnails/No.png')