+++ arcade_machine/sprites/label.py
@@ -6,27 +6,24 @@ from pygame.font import SysFont
 class Label(OrderedUpdates):
     """ Use this sprite to write text to the screen.
     """
-
-    DEFAULT_FONT = 'Arial'
     PADDING_SIZE = 5
 
     def __init__(
         self,
         text_content: str,
         text_color: str,
-        text_size: int,
         x_pos: int,
         y_pos: int,
+        font,
         background_color: str = None,
-        font_style: str = DEFAULT_FONT
     ):
         super().__init__()
         self.text_content = text_content
         self.text_color = text_color
         self.x_pos = x_pos
         self.y_pos = y_pos
+        self.font = font
         self.background_color = background_color
-        self.font = SysFont(font_style, text_size)
 
         self.label = Sprite()
         self.label.image = self.font.render(
@@ -69,3 +66,24 @@ class Label(OrderedUpdates):
             self.text_color
         )
         self.background.image.fill(self.background_color)
+
+    def redraw_label(
+            self,
+            text = None,
+            color = None,
+            font = None
+        ):
+        if text:
+            self.text_content = text
+        if color:
+            self.text_color = color
+        if font:
+            self.font = font
+        self.label.image = self.font.render(
+            self.text_content,
+            True,
+            self.text_color
+        )
+        self.label.rect = self.label.image.get_rect()
+        self.label.rect.centerx = self.x_pos
+        self.label.rect.centery = self.y_pos